adjust view of Layer
