+++ ui/src/stories/Layer/styled.ts
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2020 ZUP IT SERVICOS EM TECNOLOGIA E INOVACAO SA
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import styled from 'styled-components';
+
+const Content = styled.span`
+  color: gray;
+`;
+
+const Outside = styled.div`
+  width: 200px;
+  height: 80px;
+  background-color: lightgray;
+  border-radius: 2px;
+`;
+
+export default {
+  Content,
+  Outside,
+}
